Little more cleanup of the makefile, plus remove extra @echos
This commit removes all the echos that just say "Compiling <XXX>", this
gets lost with the rest of the compilation messages and removing them
simplifies our makefile.

Additionally we only emit instructions when make install or make all is
used. Again removing the extra @echos of the installation text simplifies
our makefile rules.