Removed server logging battles in HTML
It was time! Now with the new battle flow,
plugins will be written that will do just the same
thing (check battleclientlogs.cpp for an idea)

I at last got rid of that code that I didn't like :s

On another note, made it so stat boosts are always announced,
except sometimes silently.