In RBY, SDef and SAtk are the same thing and is called 'Special'
What we do is use the SAtk slot in the pokemon as meaning Special.
This is the convention that the client code used when implementing
the slider bar for the special stat. That means the SDef slot is left
unuesed.

For stat modifiers like amnesia that change the SDef slot,
we remap this change to the SAtk slot.