Made practical adjustments to make BattleManager work
* Changed the va_list passing, to pass it by reference. As
when you pass it in a subfunction, you pass it by value,
in the calling function the pointer wouldn't get advanced,
which would cause a bug. (everything to get the same value)
So passed it by reference everywhere.

*Removed widget memory management on the utility classes. If
a widget instantiates them, then it better use theme and free
them

* Changed the format for turn input to correct int instead of quint16

* Removed buggy 'size' property in default.css which caused
me to go crazy over debugging output (for classic theme)

* Added an example log & scene to regular battle window, so
we can still have old battle window but are able to work on the
new one at the same time.