Made the battle context accessible in the qml file
The battle data is accessible in QML through the global
data property. As Qt doesn't like templates and BattleScene
inherits from one, I had to make another class in a separate
file, BattleSceneProxy, and pass it as the main object
for the QML context. Any introducing of functionalities
will be through that proxy.