Added necessary utilities for battle window to know about stat boosts
The stat boosts, instead of a simple array, could be implemented
with a declarative list or whatever structure that allows to
notify the change of one of its elements.

Maybe in the future if we have a more advanced battle window,
it'll be that, but for now sticking to simple thing. Don't hesitate
to make the change.