Finished pokemon boxes
This is functional & working, but there are a few things not
done:

1) The drag & drop don't transfer yet from the box to the team and in
reverse. This can be done (by having the PokeButton emit a signal when
it receives mimedata, and forwarding the mimedata through the buttonholder,
for example)

2) The "Box #X Num #X" label has to be updated to either Team Slot #X
or the correct box/number

3) If you check padawan's ui, you'll see that you can access boxes from
PokeEdit. Whether or not you use a QPushButton or QToolButton or rearrange
the layout, is up to you