First part of making subgens for pokemons -- doesn't compile
The structure for pokemon subgens is like that: In the main folder we
get all the data for the whole gen. In each subfolder (Subgen X) we get
the data specific to the subgen -- i.e. the level moves of that subgen
only, if a pokemon learnt a level move in the previous subgen but doesn't
anymore then that level move isn't in the subgen folder's level moves file.

The advantage of the main folder having the data for the whole gen is that
the client doesn't need to load the data for the subgens unless needed. Moves
are heavy files, imagine having 3 times that and combining them between subgens, it could double the loading time of pokemon info.

Currently the structure are changed. MoveSet Checker is almost finished,
it only needs a loader for subgens where there's the TODO.

PokemonInfo is not finished at all.