Removed white backgrounds for gen 1/2 sprites
Used imagemagick to do this: https://gist.github.com/3102488

Some pokemon (like unown formes, or mankey) still have white
where it should be transparent, so maybe it should be fixed
- but it's really minor

The imagemagick command also works on animated gif, although
the first sprites is always present in the animation so it's
kind of bad. A fix would be appreciated for that!