Now possible to use dockwidgets in the teambuilder
I commented out real dock widget use in the team edit zone (see
the if (0) in pokeedit.cpp and teammenu.cpp), though the changes
made to the css and layout may allow it.

For example, if you put the EVs and the sprite/info in no-features
QDockWidgets, you could:
- Add the Sprite/info dockwidget with no title/no feature at the bottom
- Add the move dockwidget at the bottom (no feature/no-title)
- Split the Sprite/info dockwidget with the EV dockwidget
- Split the EV dockwidget with the level dockwidget
- Split the move dockwidget with the advanced dockwidget

With that configuration, the user can move / close the ev dockwidget,
the advanced dockwidget and the info dockwidget. Don't forget to set titles
to all the dock widgets in case the user wants to make tabs.

Also, for the no-title dockwidgets, you'll need to make it so that
QDockWidget::title-bar has a height of 0 in the css.

And it'll still have the same look, but be lots more configurable!