Spanish move names and move messages
The names were changed to match the name and case from the sixth gen games.
Lots of missing messages were added and several mistaken messages were corrected (the wrong variable was used a couple of times). There were some modifications in some messages in order to closely resemble the messages on games, too.