Battle protocol change, in prevision for zmoves going through Taunt
Now whether a zmove or not is possible to be selected in the client's UI is decided by the server, instead of the client making that deduction itself.

It is necessary for when some moves are unable to be selected but their corresponding zmove is. I don't know the exact mechanics of that, whether it is that zmoves are always being able to be selected unless there is a lock on another move, or not.

In any case, better put it in the protocol *now* rather than after gen7 rollout, where it would break some gen7 clients in doubles/triples.